Moved everything to examples folder, split strategy into 2 steps
