+++ arroyo/examples/__init__.py
@@ -1,0 +1,3 @@
+# from .transform_and_produce import HashPasswordAndProduceStrategyFactory
+
+# __all__ = ["HashPasswordAndProduceStrategyFactory"]
