+++ arroyo/examples/transform_and_produce/hash_password_strategy.py
@@ -1,0 +1,65 @@
+import hashlib
+import json
+import logging
+from typing import Optional
+
+from arroyo.backends.kafka.consumer import KafkaPayload
+from arroyo.processing.strategies.abstract import (
+    ProcessingStrategy,
+    ProcessingStrategy as ProcessingStep,
+)
+
+from arroyo.types import Message
+
+logger = logging.getLogger(__name__)
+
+
+class HashPasswordStrategy(ProcessingStrategy[KafkaPayload]):
+    """
+    Hash Password Step.
+    """
+
+    def __init__(self, next_step: ProcessingStep[KafkaPayload]) -> None:
+        self.__next_step = next_step
+        self.__closed = False
+
+    def poll(self) -> None:
+        self.__next_step.poll()
+
+    def submit(self, message: Message[KafkaPayload]) -> None:
+        assert not self.__closed
+
+        auth = json.loads(message.payload.value)
+        hashed = self._hash_password(auth["password"])
+        data = json.dumps({"username": auth["username"], "password": hashed}).encode(
+            "utf-8"
+        )
+
+        payload = KafkaPayload(
+            message.payload.key, data, headers=message.payload.headers
+        )
+
+        self.__next_step.submit(
+            Message(
+                message.partition,
+                message.offset,
+                payload,
+                message.timestamp,
+            )
+        )
+
+    def _hash_password(self, password: str) -> str:
+        return hashlib.sha256(password.encode("utf-8")).hexdigest()
+
+    def close(self) -> None:
+        self.__closed = True
+
+    def terminate(self) -> None:
+        self.close()
+
+        logger.debug("Terminating %r...", self.__next_step)
+        self.__next_step.terminate()
+
+    def join(self, timeout: Optional[float] = None) -> None:
+        self.__next_step.close()
+        self.__next_step.join(timeout)
